Refactor Viewcontroller to CardViewController
change variable name and divide drawCard function
make static variable for card ratio
